move HSDS test to resource extraction tests
